requirements.txt
pandas
pycryptodomex
requests
flask
flask_cors
utils
